Refactor identity calls to use module with cache.
