+++ client/scripts/controllers/chain/substrate/identities.ts
@@ -1,0 +1,274 @@
+import { IApp } from 'state';
+import { StorageModule } from 'models';
+import { ProposalStore } from 'stores';
+import { SubstrateCoin } from 'adapters/chain/substrate/types';
+import {
+  Call,
+  Registration,
+  AccountId,
+  RegistrarInfo,
+  IdentityJudgement,
+  IdentityFields
+} from '@polkadot/types/interfaces';
+import { Codec } from '@polkadot/types/types';
+import { Vec, Option, Data } from '@polkadot/types';
+import { Unsubscribable, Observable, of } from 'rxjs';
+import { first, takeWhile, map } from 'rxjs/operators';
+import { ApiRx } from '@polkadot/api';
+import BN from 'bn.js';
+import SubstrateChain from './shared';
+import SubstrateAccounts, { SubstrateAccount } from './account';
+import SubstrateIdentity from './identity';
+
+class SubstrateIdentityStore extends ProposalStore<SubstrateIdentity> { }
+
+export type SuperCodec = [ AccountId, Data ] & Codec;
+export type IdentityInfoProps = {
+  display?: any;
+  legal?: any;
+  web?: any;
+  riot?: any;
+  email?: any;
+  pgpFingerprint?: any;
+  twitter?: any;
+  image?: any;
+  additional: any[];
+};
+
+class SubstrateIdentities implements StorageModule {
+  private _initialized: boolean = false;
+  public get initialized() { return this._initialized; }
+
+  private _store: SubstrateIdentityStore = new SubstrateIdentityStore();
+  public get store() { return this._store; }
+
+  private _Chain: SubstrateChain;
+  private _Accounts: SubstrateAccounts;
+
+  private _registrarSubscription: Unsubscribable;
+  private _registrars: Array<RegistrarInfo | null>; // with gaps
+  public get registrars() { return this._registrars; }
+
+  private _fieldDeposit: SubstrateCoin;
+  private _basicDeposit: SubstrateCoin;
+  private _subAcctDeposit: SubstrateCoin;
+  private _maxSubAccts: number;
+  private _maxAddlFields: number;
+  public get fieldDeposit() { return this._fieldDeposit; }
+  public get basicDeposit() { return this._basicDeposit; }
+  public get subAcctDeposit() { return this._subAcctDeposit; }
+  public get maxSubAccts() { return this._maxSubAccts; }
+  public get maxAddlFields() { return this._maxAddlFields; }
+
+  private _app: IApp;
+  public get app() { return this._app; }
+
+  constructor(app: IApp) {
+    this._app = app;
+  }
+
+  public deinit() {
+    if (this._registrarSubscription) {
+      this._registrarSubscription.unsubscribe();
+    }
+    this._initialized = false;
+    this.store.clear();
+  }
+
+  // given an account, fetch the corresponding identity (works on sub-accounts)
+  public get(who: SubstrateAccount): Observable<SubstrateIdentity> {
+    // check immediately if we have the id
+    const existingIdentity = this.store.getByIdentifier(who.address);
+    if (existingIdentity) {
+      return of(existingIdentity);
+    }
+
+    // NOTE: the block of code below is commented out, because we do not necessarily want to
+    //   resolve sub-addresses to their super-identities -- this is a decision requiring discussion.
+    // check for a super-identity (maybe they passed in a sub address)
+    let acctToFetch = who;
+    // const superId = await this._Chain.query(
+    //   (api: ApiRx) => api.query.identity.superOf<Option<SuperCodec>>(who.address).pipe(first())
+    // ).toPromise();
+    // if (superId.isSome) {
+    //   const superAcct = this._Accounts.get(superId.unwrap()[0].toString());
+    //   const existingSuperId = this.store.getByIdentifier(superAcct.address);
+    //   if (existingSuperId) {
+    //     return of(existingSuperId);
+    //   }
+    //   acctToFetch = superAcct;
+    // }
+
+    // check on chain for registration we haven't seen yet & wait for it to appear
+    // if we have a super address to fetch, grab that instead
+    return this._Chain.query(
+      (api: ApiRx) => api.query.identity.identityOf<Option<Registration>>(acctToFetch.address)
+    ).pipe(
+      takeWhile((id) => !id.isSome, true),
+      map((id) => id.isSome
+        ? new SubstrateIdentity(this._Chain, this._Accounts, this, acctToFetch, id.unwrap())
+        : null),
+    );
+  }
+
+  public init(ChainInfo: SubstrateChain, Accounts: SubstrateAccounts): Promise<void> {
+    this._Chain = ChainInfo;
+    this._Accounts = Accounts;
+    return new Promise((resolve) => {
+      this._Chain.api.pipe(first()).subscribe((api: ApiRx) => {
+        // init consts
+        // XXX: for now, these aren't exposed in the Rust code. Which means the module isn't
+        //   ready for use. Since the module isn't ready, the module is disabled on Substrate
+        //   and Edgeware.
+        //
+        // pub const BasicDeposit: Balance = 10 * DOLLARS;       // 258 bytes on-chain
+        // pub const FieldDeposit: Balance = 250 * CENTS;        // 66 bytes on-chain
+        // pub const SubAccountDeposit: Balance = 2 * DOLLARS;   // 53 bytes on-chain
+        // pub const MaxSubAccounts: u32 = 100;
+        // pub const MaxAdditionalFields: u32 = 100;
+
+        // this._basicDeposit = this._Chain.coins(api.consts.identity.basicDeposit as BalanceOf);
+        // this._fieldDeposit = this._Chain.coins(api.consts.identity.fieldDeposit as BalanceOf);
+        // this._subAcctDeposit = this._Chain.coins(api.consts.identity.subAccountDeposit as BalanceOf);
+        // this._maxSubAccts = +api.consts.identity.maxSubAccounts;
+        // this._maxAddlFields = +api.consts.identity.maxAdditionalFields;
+        if (!this._basicDeposit) this._basicDeposit = this._Chain.coins(10, true);
+        if (!this._fieldDeposit) this._fieldDeposit = this._Chain.coins(2.5, true);
+        if (!this._subAcctDeposit) this._subAcctDeposit = this._Chain.coins(2, true);
+        if (!this._maxSubAccts) this._maxSubAccts = 100;
+        if (!this._maxAddlFields) this._maxAddlFields = 100;
+
+        // kick off long-running subscription to registrars
+        this._registrarSubscription = this._Chain.query(
+          (apiRx: ApiRx) => apiRx.query.identity.registrars()
+        ).subscribe((rs: Vec<Option<RegistrarInfo>>) => {
+          this._registrars = rs.map((r) => r.unwrapOr(null));
+          if (!this._initialized) {
+            this._initialized = true;
+            resolve();
+          }
+        });
+      });
+    });
+  }
+
+  // TRANSACTIONS
+  // TODO: add helper for mashalling substrate Data fields
+  public async setIdentityTx(who: SubstrateAccount, data: IdentityInfoProps) {
+    const info = this._Chain.createType('IdentityInfo', data);
+    if (info.additional.length > this.maxAddlFields) {
+      throw new Error('too many additional fields');
+    }
+
+    // compute the basic required balance for the registration
+    let requiredBalance = this.basicDeposit.add(this.fieldDeposit.muln(info.additional.length));
+
+    // compare with preexisting deposit from old registration, if exists
+    const oldId = this.store.getByIdentifier(who.address);
+    if (oldId && oldId.deposit.lt(requiredBalance)) {
+      requiredBalance = requiredBalance.sub(oldId.deposit);
+    } else if (oldId && oldId.deposit.gte(requiredBalance)) {
+      requiredBalance = new BN(0);
+    }
+
+    // verify the account has sufficient funds based on above computation
+    const txFunc = (api: ApiRx) => api.tx.identity.setIdentity(info);
+    if (!(await this._Chain.canPayFee(who, txFunc, this._Chain.coins(requiredBalance)))) {
+      throw new Error('insufficient funds');
+    }
+    return this._Chain.createTXModalData(
+      who,
+      txFunc,
+      'setIdentity',
+      `${who.address} registers identity ${info.display.toString()}`
+    );
+  }
+
+  public setRegistrarFeeTx(who: SubstrateAccount, regIdx: number, fee: SubstrateCoin) {
+    if (!this.registrars[regIdx] || !this.registrars[regIdx].account) {
+      throw new Error('invalid registrar');
+    }
+    if (this.registrars[regIdx].account.toString() !== who.address) {
+      throw new Error('invalid account');
+    }
+    return this._Chain.createTXModalData(
+      who,
+      (api: ApiRx) => api.tx.identity.setFee(regIdx, fee),
+      'setFee',
+      `registrar ${regIdx} updates fee to ${fee.format(true)}`,
+    );
+  }
+
+  public setRegistrarAccountTx(who: SubstrateAccount, regIdx: number, newAcct: SubstrateAccount) {
+    if (!this.registrars[regIdx] || !this.registrars[regIdx].account) {
+      throw new Error('invalid registrar');
+    }
+    if (this.registrars[regIdx].account.toString() !== who.address) {
+      throw new Error('invalid account');
+    }
+    return this._Chain.createTXModalData(
+      who,
+      (api: ApiRx) => api.tx.identity.setAccountId(regIdx, newAcct.address),
+      'setAccountId',
+      `registrar ${regIdx} updates account to ${newAcct.address}`,
+    );
+  }
+
+  public setRegistrarFieldsTx(who: SubstrateAccount, regIdx: number, fields: IdentityFields) {
+    if (!this.registrars[regIdx] || !this.registrars[regIdx].account) {
+      throw new Error('invalid registrar');
+    }
+    if (this.registrars[regIdx].account.toString() !== who.address) {
+      throw new Error('invalid account');
+    }
+    return this._Chain.createTXModalData(
+      who,
+      (api: ApiRx) => api.tx.identity.setFields(regIdx, fields),
+      'setFee',
+      `registrar ${regIdx} updates fields`,
+    );
+  }
+
+  public providejudgementTx(
+    who: SubstrateAccount,
+    regIdx: number,
+    target: SubstrateIdentity,
+    judgement: IdentityJudgement
+  ) {
+    if (!this.registrars[regIdx] || !this.registrars[regIdx].account) {
+      throw new Error('invalid registrar');
+    }
+    if (this.registrars[regIdx].account.toString() !== who.address) {
+      throw new Error('invalid account');
+    }
+    if (!target.exists) {
+      throw new Error('target identity does not exist');
+    }
+    if (judgement.isFeePaid) {
+      throw new Error('invalid judgement');
+    }
+    return this._Chain.createTXModalData(
+      who,
+      (api: ApiRx) => api.tx.identity.provideJudgement(regIdx, target.account.address, judgement),
+      'providejudgement',
+      `registrar ${regIdx} provides judgement for identity ${target.username}`,
+    );
+  }
+
+  // requires RegistrarOrigin or Root!
+  public addRegistrarMethod(account: SubstrateAccount): Call {
+    const func = this._Chain.getTxMethod('identity', 'addRegistrar');
+    return func(account.address).method;
+  }
+
+  // requires ForceOrigin or Root!
+  public killIdentityMethod(target: SubstrateIdentity): Call {
+    if (!target.exists) {
+      throw new Error('target identity does not exist');
+    }
+    const func = this._Chain.getTxMethod('identity', 'killIdentity');
+    return func(target.account.address).method;
+  }
+}
+
+export default SubstrateIdentities;
