+++ client/scripts/views/modals/edit_identity_modal.ts
@@ -4,18 +4,16 @@ import m from 'mithril';
 import $ from 'jquery';
 import app from 'state';
 
-import { createType } from '@polkadot/types/create';
 import { IdentityInfo } from '@polkadot/types/interfaces';
 import { Data } from '@polkadot/types/primitive';
 import { u8aToString } from '@polkadot/util';
 
 import CharacterLimitedTextInput from '../components/widgets/character_limited_text_input';
 import { createTXModal } from './tx_signing_modal';
-import { Account } from 'models';
 import { SubstrateAccount } from '../../controllers/chain/substrate/account';
 import AvatarUpload from '../components/avatar_upload';
 import Substrate from '../../controllers/chain/substrate/main';
-import { IdentityInfoProps } from '../../controllers/chain/substrate/identity';
+import { IdentityInfoProps } from '../../controllers/chain/substrate/identities';
 
 interface IAttrs {
   currentIdentity?: IdentityInfo;
